remove unused files and renaming some components
