--- src/components/dialog/edit-dialog.jsx
@@ -1,11 +1,0 @@
-import { styled } from '@mui/material/styles';
-import Dialog from '@mui/material/Dialog';
-
-const BootstrapDialog = styled(Dialog)(({ theme }) => ({
-  '& .MuiDialogContent-root': {
-    padding: theme.spacing(2),
-  },
-  '& .MuiDialogActions-root': {
-    padding: theme.spacing(1),
-  },
-}));
